move stream
