+++ pkg/lifecycle/stream/base.go
