+++ pkg/lifecycle/stream/destination.go
