+++ pkg/lifecycle/stream/destination_test.go
