+++ pkg/lifecycle/stream/doc.go
