+++ pkg/lifecycle/stream/fanin.go
