+++ pkg/lifecycle/stream/fanin_select.go
