+++ pkg/lifecycle/stream/fanout.go
