+++ pkg/lifecycle/stream/fanout_test.go
