+++ pkg/lifecycle/stream/message.go
