+++ pkg/lifecycle/stream/message_test.go
