+++ pkg/lifecycle/stream/metrics.go
