+++ pkg/lifecycle/stream/mock/destination.go
