+++ pkg/lifecycle/stream/mock/processor.go
