+++ pkg/lifecycle/stream/mock/source.go
