+++ pkg/lifecycle/stream/node.go
