+++ pkg/lifecycle/stream/parallel.go
