+++ pkg/lifecycle/stream/parallel_test.go
