+++ pkg/lifecycle/stream/processor.go
