+++ pkg/lifecycle/stream/processor_test.go
