+++ pkg/lifecycle/stream/source_test.go
