+++ pkg/lifecycle/stream/stream_test.go
