+++ pkg/orchestrator/orchestrator.go
@@ -128,7 +128,7 @@ type ProcessorPluginService interface {
 }
 
 type LifecycleService interface {
-	Start(ctx context.Context, connFetcher lifecycle.ConnectorFetcher, procService lifecycle.ProcessorService, pluginFetcher lifecycle.PluginDispenserFetcher, pipelineID string) error
+	Start(ctx context.Context, connFetcher lifecycle.ConnectorFetcher, procService lifecycle.ProcessorService, pluginFetcher lifecycle.PluginDispenserFetcher, lifecycleService lifecycle.Service, pipelineID string) error
 	// Stop initiates a stop of the given pipeline. The method does not wait for
 	// the pipeline (and its nodes) to actually stop.
 	// When force is false the pipeline will try to stop gracefully and drain
