+++ pkg/provisioning/service.go
@@ -40,6 +40,7 @@ type Service struct {
 	connectorService       ConnectorService
 	processorService       ProcessorService
 	connectorPluginService ConnectorPluginService
+	lifecycleService       LifecycleService
 	pipelinesPath          string
 }
 
@@ -254,8 +255,7 @@ func (s *Service) provisionPipeline(ctx context.Context, cfg config.Pipeline) er
 	if cfg.Status == config.StatusRunning {
 		// TODO set status and let the pipeline service start it
 
-		// TODO: Use the lifecycle service (not the pipeline service) to start the pipeline
-		err := s.pipelineService.Start(ctx, s.connectorService, s.processorService, s.connectorPluginService, cfg.ID)
+		err := s.lifecycleService.Start(ctx, s.connectorService, s.processorService, s.connectorPluginService, cfg.ID)
 		if err != nil {
 			return cerrors.Errorf("could not start the pipeline %q: %w", cfg.ID, err)
 		}
